+++ desktop/main/launcher.ts
@@ -38,8 +38,9 @@ export function createLauncherWindow(): BrowserWindow {
       : { frame: false }),
 
     webPreferences: {
-      nodeIntegration: true,
-      contextIsolation: false,
+      nodeIntegration: false,
+      contextIsolation: true,
+      preload: path.join(__dirname, 'preload.js'),
     },
   });